move ExperimentWriter to logging folder
