+++ all/experiments/__init__.py
@@ -3,8 +3,6 @@ from .experiment import Experiment
 from .single_env_experiment import SingleEnvExperiment
 from .parallel_env_experiment import ParallelEnvExperiment
 from .multiagent_env_experiment import MultiagentEnvExperiment
-from .writer import ExperimentWriter
-from .writer import CometWriter
 from .plots import plot_returns_100
 from .slurm import SlurmExperiment
 from .watch import watch, load_and_watch
@@ -16,8 +14,7 @@ __all__ = [
     "ParallelEnvExperiment",
     "MultiagentEnvExperiment",
     "SlurmExperiment",
-    "ExperimentWriter",
-    "CometWriter",
     "watch",
     "load_and_watch",
+    "plot_returns_100"
 ]
