+++ all/logging/dummy.py
@@ -1,0 +1,23 @@
+from ._writer import Writer
+
+
+class DummyWriter(Writer):
+    '''A default Logger object that performs no logging and has no side effects.'''
+
+    def add_loss(self, name, value, step="frame"):
+        pass
+
+    def add_evaluation(self, name, value, step="frame"):
+        pass
+
+    def add_scalar(self, name, value, step="frame"):
+        pass
+
+    def add_schedule(self, name, value, step="frame"):
+        pass
+
+    def add_summary(self, name, mean, std, step="frame"):
+        pass
+
+    def close(self):
+        pass
